Refactor platform API imports to use the new module structure and add Electron-specific Platform API implementation
