+++ apps/electron/src/renderer/platform-api/hooks/use-platform-api.ts
